Revert "first attempt at network batching rewrite"
This reverts commit 7a7870a77f202500ff11dbafe6982c8a39be5cff.
