+++ Wikipedia/Code/ReadingListsController.swift
@@ -5,6 +5,8 @@ internal let WMFReadingListDefaultListEnabledKey = "WMFReadingListDefaultListEna
 
 internal let WMFReadingListUpdateKey = "WMFReadingListUpdateKey"
 
+internal let WMFReadingListBatchRequestLimit = 8 // currently this waits until all requests are done before firing off new ones, could be optimized to add new requests as old ones finish
+
 internal let WMFReadingListBatchSizePerRequestLimit = 500
 
 internal let WMFReadingListCoreDataBatchSize = 500
