+++ Wikipedia/Code/Session.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-@objc(WMFSession) public class Session: NSObject {
+public class Session {
     public struct Request {
         public enum Method {
             case get
@@ -31,15 +31,9 @@ import Foundation
 
     }
 
-    @objc public static let shared = Session()
+    public static let shared = Session()
 
     fileprivate let session = URLSession.shared
-    
-    private lazy var queue: OperationQueue = {
-        let queue = OperationQueue()
-        queue.maxConcurrentOperationCount = 16
-        return queue
-    }()
 
     public func mediaWikiAPITask(host: String, scheme: String = "https", method: Session.Request.Method = .get, queryParameters: [String: Any]? = nil, bodyParameters: Any? = nil, completionHandler: @escaping ([String: Any]?, HTTPURLResponse?, Error?) -> Swift.Void) -> URLSessionDataTask? {
         return jsonDictionaryTask(host: host, scheme: scheme, method: method, path: WMFAPIPath, queryParameters: queryParameters, bodyParameters: bodyParameters, bodyEncoding: .form, completionHandler: completionHandler)
@@ -108,7 +102,7 @@ import Foundation
         guard let request = request(host: host, scheme: scheme, method: method, path: path, queryParameters: queryParameters, bodyParameters: bodyParameters, bodyEncoding: bodyEncoding) else {
             return nil
         }
-        return jsonDictionaryTask(with: request, completionHandler: { (result, response, error) in
+        return session.wmf_jsonDictionaryTask(with: request, completionHandler: { (result, response, error) in
             completionHandler(result, response as? HTTPURLResponse, error)
         })
     }
@@ -168,83 +162,5 @@ import Foundation
                 handleErrorResponse()
             }
         })
-    }
-    
-    
-    public func jsonDictionaryTask(with request: URLRequest, completionHandler: @escaping ([String: Any]?, URLResponse?, Error?) -> Swift.Void) -> URLSessionDataTask {
-        return session.dataTask(with: request, completionHandler: { (data, response, error) in
-            guard let data = data else {
-                completionHandler(nil, response, error)
-                return
-            }
-            do {
-                guard data.count > 0, let responseObject = try JSONSerialization.jsonObject(with: data, options: []) as? [String: Any] else {
-                    completionHandler(nil, response, nil)
-                    return
-                }
-                completionHandler(responseObject, response, nil)
-            } catch let error {
-                DDLogError("Error parsing JSON: \(error)")
-                completionHandler(nil, response, error)
-            }
-        })
-    }
-    
-    public func summaryTask(with articleURL: URL, completionHandler: @escaping ([String: Any]?, URLResponse?, Error?) -> Swift.Void) -> URLSessionDataTask? {
-        guard let siteURL = articleURL.wmf_site, let title = articleURL.wmf_titleWithUnderscores else {
-            return nil
-        }
-        
-        let encodedTitle = title.addingPercentEncoding(withAllowedCharacters: CharacterSet.wmf_urlPathComponentAllowed) ?? title
-        let percentEncodedPath = NSString.path(withComponents: ["/api", "rest_v1", "page", "summary", encodedTitle])
-        
-        guard var components = URLComponents(url: siteURL, resolvingAgainstBaseURL: false) else {
-            return nil
-        }
-        components.percentEncodedPath = percentEncodedPath
-        guard let summaryURL = components.url else {
-            return nil
-        }
-        
-        var request = URLRequest(url: summaryURL)
-        //The accept profile is case sensitive https://gerrit.wikimedia.org/r/#/c/356429/
-        request.setValue("application/json; charset=utf-8; profile=\"https://www.mediawiki.org/wiki/Specs/Summary/1.1.2\"", forHTTPHeaderField: "Accept")
-        return jsonDictionaryTask(with: request, completionHandler: completionHandler)
-    }
-    
-    @objc(fetchSummaryWithArticleURL:completionHandler:)
-    public func fetchSummary(with articleURL: URL, completionHandler: @escaping ([String: Any]?, URLResponse?, Error?) -> Swift.Void) {
-        guard let task = summaryTask(with: articleURL, completionHandler: completionHandler) else {
-            completionHandler(nil, nil, NSError.wmf_error(with: .invalidRequestParameters))
-            return
-        }
-        let operation = URLSessionTaskOperation(task: task)
-        queue.addOperation(operation)
-    }
-    
-    public func fetchArticleSummaryResponsesForArticles(withURLs articleURLs: [URL], completion: @escaping ([String: [String: Any]]) -> Void) {
-        let queue = DispatchQueue(label: "ArticleSummaryFetch-" + UUID().uuidString)
-        let taskGroup = WMFTaskGroup()
-        var summaryResponses: [String: [String: Any]] = [:]
-        for articleURL in articleURLs {
-            guard let key = articleURL.wmf_articleDatabaseKey else {
-                continue
-            }
-            taskGroup.enter()
-            fetchSummary(with: articleURL, completionHandler: { (responseObject, response, error) in
-                guard let responseObject = responseObject else {
-                    taskGroup.leave()
-                    return
-                }
-                queue.async {
-                    summaryResponses[key] = responseObject
-                    taskGroup.leave()
-                }
-            })
-        }
-        
-        taskGroup.waitInBackgroundAndNotify(on: queue) {
-            completion(summaryResponses)
-        }
     }
 }
